add rules and add a src map
