+++ App.tsx
@@ -1,7 +1,7 @@
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
-import RootNavigator from './Navigation/RootNavigator';
+import RootNavigator from './src/Navigation/RootNavigator';
 
 export default function App() {
   return (
