--- Screens/RegisterScreen.tsx
@@ -1,25 +1,0 @@
-import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { StyleSheet, Text, View } from 'react-native';
-import { RootStackParamList } from '../Navigation/RootNavigator';
-
-type Props = NativeStackScreenProps<RootStackParamList, 'RegisterScreen'>;
-
-export default function RegisterScreen({ navigation }: Props) {
-  return (
-    <View style={styles.container}>
-      <Text>Registrera användare</Text>
-      <Text>Användarnamn</Text>
-      <Text>Lösenord</Text>
-      <Text>Lösenord igen</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
