+++ package.json
@@ -7,8 +7,9 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "lint": "eslint .",
-    "format": "prettier --check ./src"
+    "lint": "eslint ./src",
+    "format": "prettier ./ --check",
+    "prettier": "npm run format -- --write"
   },
   "dependencies": {
     "@react-navigation/native": "^6.0.13",
