+++ src/Screens/ProfileScreen.tsx
@@ -1,15 +1,17 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import * as React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { RootStackParamList } from '../Navigation/RootNavigator';
 
-type Props = NativeStackScreenProps<RootStackParamList, 'HomeScreen'>;
+type Props = NativeStackScreenProps<RootStackParamList, 'ProfileScreen'>;
 
-export default function HomeScreen({ navigation }: Props) {
+export default function ProfileScreen({ navigation }: Props) {
   return (
     <View style={styles.container}>
-      <Text>Chores to be done</Text>
-      <Text>Add chore</Text>
-      <Text>Edit chore</Text>
+      <Text>Profile screen</Text>
+      <Text>Name</Text>
+      <Text>Avatar</Text>
+      <Text>Knapp: Skapa</Text>
     </View>
   );
 }
