+++ src/Screens/RegisterScreen.tsx
@@ -1,17 +1,17 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import * as React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { RootStackParamList } from '../Navigation/RootNavigator';
 
-type Props = NativeStackScreenProps<RootStackParamList, 'ChoreScreen'>;
+type Props = NativeStackScreenProps<RootStackParamList, 'RegisterScreen'>;
 
-export default function ChoreScreen({ navigation }: Props) {
+export default function RegisterScreen({ navigation }: Props) {
   return (
     <View style={styles.container}>
-      <Text>TextInput: Titel</Text>
-      <Text>TextInput: Beskrivning</Text>
-      <Text>Återkommer: var blabla dag </Text>
-      <Text>Värde: 1-2-4-6-8 </Text>
-      <Text>SKALL HA EDIT </Text>
+      <Text>Registrera användare</Text>
+      <Text>Användarnamn</Text>
+      <Text>Lösenord</Text>
+      <Text>Lösenord igen</Text>
     </View>
   );
 }
